requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
Jinja2>=3.1
itsdangerous>=2.1
Pillow>=10.0
google-genai>=0.3.0
python-dotenv>=1.0.0
rembg[cpu]>=2.0.56
psycopg[binary]>=3.1.18
celery>=5.3.6
redis>=5.0.8
PyPDF2>=3.0.0
